Give hostdb and allowance field which gets set alongside weightFunc
Renamed calculateHostWeight to weightFunc as a field in the hostdb, and
then put alongside it an allowance field so that the hostdb knows what
allowance was used to generate the weightFunc.

EstimateHostScore now takes an allowance as input, so that the host
score can be estimated for any given allowance, as the host scores are
allowance dependent.

ScoreBreakdown on the other hand does not take an allowance as input,
instead it uses the allowance that is currently active in the hostdb,
giving the caller an accurate breakdown of the host score that reflects
the actual score of the host in the hostdb.