Add Unlock Hash From String Utility Function
We should add the ScanAddress and UnlockHashFromAddrStr functions to `types/encoding.go` so that we can access these funstions in `types/constants.go`. Doing so will allow us to supply the unlockHashes in string form which is the form that most users will be familiar with. This in turn will help with ease of documenting code and writing tutorials.