Update the 'close' methods of the SiaFileSetEntry
The previous implementation of 'Close' for the SiaFileSetEntry was not
safe, as the implementation had the SiaFileSetEntry making modifications
to the SiaFileSet without calling out to a method of the SiaFileSet.

The SiaFileSet is the superior object in this case, which means that the
SiaFileSet needs to be controlling all of the state transitions. For the
case of closing siaFileEntries, the previous implementation assumed
incorrectly that the SiaFileSet had the same entry in the SiaFileMap as
the entry that had been closed. This caused a rare but significant (file
corruption) bug.

The code has been restructured now so that the SiaFileSet performs the
close action, and manages its own maps, accounting for the fact that the
entry may have changed.