remove unneeded wrapper, improve exported docstring
The docstring of an exported function is the only thing that someone is
going to see in a godoc, they will not be able to see the docstring on
the unexported function. A docstring that says 'this is a wrapper of the
unexported version of this function' is therefore not very helpful.

Further, there's no need to have an unexported version of a function if
that's not being used anywhere, and it adds unecessary indirection to
the code.