use more hosts when forming contracts
A significant fraction of the host network is currently unable to form
contracts due to either the wallet being locked, due to their collateral
budget being drained (a bug that was fixed in v1.4.0), and due to the
wallet being empty.

v1.4.0 should help with a lot of these issues, but in the meantime we
need to be requesting more hosts from the hostdb so that we can tolerate
more failures.