remove health check conditional
previously we would move on to a new directory if the directory we were
working on was reasonably healthy, to ensure that we were always looking
that the lowest health things.

A new design for the uploading code fixes this concern entirely in a
different way, removing this to reduce complexity.