Add Genesis Siacoin Allocation
The Sia blockchain does not set a genesis siacoin allocation.
However, having the constants in place to support a genesis siacoin
allocation would make it easier for genesis forks of Sia to set up
an airdrop. It would also simplify documentation on how to execute
an airdrop when starting a genesis fork of Sia.

Additionally, the Genesis block should not be limited to having all
Siacoin and Siafund outputs in the same transaction. Doing so causes
confusion as there is no limitation in the types.constants.go file
but there is elsewhere in the code. The end result is that without
this change there is an undocumented constraint on how the Genesis
block should be structured.