remove fileRepairInterval code
The file repair interval code would ingore files that had been marked as
having been seen/viewed/updated recently. This would potentially cause
the health loop and repair loop to skip over a file that needed work, or
at least ignore the file for up to 5 mintues.

This code adds significant complexity to the design because it means you
have to watch out for scenarios where a file is just going to be
ignored.

The new design for the repair loop should mitigate the original issue
that the file repair interval intended to solve without adding this
complexity.