perform cleanup related to the buildUnfinishedChunks method
The code for processing unfinished chunks was incorrectly iterating
through an array that could have nil chunks at the tail. The way that
the array is constructed has been adjusted to prevent panics in the
future.

Additional cleanup was performed by returning an error in the even that
a chunk could not be built, instead of just returning a nil chunk.
Switching to returning an error caught an additional bug in a separate
piece of code which had assumed that a non-nil chunk would always be
returned.

Two anti-patters were discovered in this code. The first anti-pattern
was fully alloacating an array and then setting each element one object
at a time. This is an anti-pattern because if at any point the process
of filling out the array is interrupted (or objects are skipped), the
array will have an invisible tail of zero or nil objects in it. This
anti-pattern directly caused a bug in this case.

The second anti-pattern was having a function that could return a 'nil'
object upon experiencing an error, and not having that function return
an error. This resulted in a bug that was not caught by testing, and had
the pattern not been changed to be explicit, this bug could have
manifested in production.