Maintain persistent set of nodes connected to
Use the persist module to save mapping from IP address
to timestamp of the last successful connection for each node
scanned.

The node scanner now either starts its scan using this persistent set.
If it is being used for the first time, then it starts with the
bootstrap set instead.