move compat code into persist_compat.go
A couple of structs and functions are still defined in files.go that are
misleading because they are only used in the compat process to upgrade
an old file format. Some documentation has been added to express this
fact, and all of the compat-only code has been moved to
persist_compat.go.