Update release.sh for new release process.
Save old release.sh for use in nightly builds in the future.

The updated release.sh now does no signing. It hashes each of the
binaries and appends the hashes to a single file. Copies of that file
are put into each zipped package. Before releasing, a signature of the
hash file must be added to each zipped package.