Add reproducible build script
release-builder.Dockerfile defines a basic Docker image that pulls a
specified branch from this repository and builds it using release.sh and
the specified version.

build-in-docker.sh creates a Docker container using
release-builder.Dockerfile, copies out the results, and deletes the
container. The script also prints out the SHA256SUM file from the Docker
builds for quick verification.

This enables reproducible builds because the outputs of
`./build-docker.sh $branch $version` should be deterministic.